新增全局異常處理器功能、更新openapi模組的版本

+++ src/main/java/com/oovetest/webDemo/series/dto/SeriesResponse.java
@@ -8,7 +8,9 @@ import lombok.Data;
 public class SeriesResponse {
     private Long id;
     private String title;
-    private Integer bookCount; 
+    private Long bookCount; 
     private SeriesStatus status;
     private Long authorId;
+
+    
 }
